distsql: move span resolver to distsqlplan
